Setup endpoint to refresh votes
This creates a /reinitialize endpoint (DELETE verb) on the twith-chat
service so other services can cause a reset of the votes.

csgo-poker can use this today by hitting it after it fetches votes when
it sees a round start on the console

When something can consume gsi, that service can hit reinitialize
whenever it sees the start of freezetime.